// src/templates/formEditarAvalLivro/types/FormData.ts
export interface FormData {
  nomeAvaliado: string;
  matricula: number;
  idAvaliador: number;
  nomeAvaliador: string;
  numeroTurma: number;
  dataEHora: string;
  dataValidacao: string;
  estetica: string;
  dignidade: string;
  clareza: string;
  plagio: string;
  observacao: string;
  conceito: string;
  opniao: string;
  sociedade: string;
  plagioParcial: string;
  relevantes: string;
  redacao: string;
  portugues: string;
  historiaObservacao: string;
  historiaRelatorio: string;
}